+++ great_expectations/data_context/data_context/base_data_context.py
@@ -48,17 +48,9 @@ from great_expectations.data_context.util import (
 )
 from great_expectations.datasource import LegacyDatasource
 from great_expectations.datasource.new_datasource import BaseDatasource, Datasource
-from great_expectations.profile.basic_dataset_profiler import BasicDatasetProfiler
 from great_expectations.render.renderer.site_builder import SiteBuilder
 from great_expectations.rule_based_profiler import RuleBasedProfiler
 from great_expectations.validator.validator import Validator
-
-try:
-    from sqlalchemy.exc import SQLAlchemyError
-except ImportError:
-    # We'll redefine this error in code below to catch ProfilerError, which is caught above, so SA errors will
-    # just fall through
-    SQLAlchemyError = ge_exceptions.ProfilerError
 
 logger = logging.getLogger(__name__)
 
@@ -174,10 +166,6 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
     --ge-feature-maturity-info--
     """
 
-    PROFILING_ERROR_CODE_TOO_MANY_DATA_ASSETS = 2
-    PROFILING_ERROR_CODE_SPECIFIED_DATA_ASSETS_NOT_FOUND = 3
-    PROFILING_ERROR_CODE_NO_BATCH_KWARGS_GENERATORS_FOUND = 4
-    PROFILING_ERROR_CODE_MULTIPLE_BATCH_KWARGS_GENERATORS_FOUND = 5
     UNCOMMITTED_DIRECTORIES = ["data_docs", "validations"]
     GE_UNCOMMITTED_DIR = "uncommitted"
     BASE_DIRECTORIES = [
@@ -777,228 +765,6 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
             logger.debug("No data_docs_config found. No site(s) built.")
 
         return index_page_locator_infos
-
-    def profile_datasource(  # noqa: C901 - complexity 25
-        self,
-        datasource_name,
-        batch_kwargs_generator_name=None,
-        data_assets=None,
-        max_data_assets=20,
-        profile_all_data_assets=True,
-        profiler=BasicDatasetProfiler,
-        profiler_configuration=None,
-        dry_run=False,
-        run_id=None,
-        additional_batch_kwargs=None,
-        run_name=None,
-        run_time=None,
-    ):
-        """Profile the named datasource using the named profiler.
-
-        Args:
-            datasource_name: the name of the datasource for which to profile data_assets
-            batch_kwargs_generator_name: the name of the batch kwargs generator to use to get batches
-            data_assets: list of data asset names to profile
-            max_data_assets: if the number of data assets the batch kwargs generator yields is greater than this max_data_assets,
-                profile_all_data_assets=True is required to profile all
-            profile_all_data_assets: when True, all data assets are profiled, regardless of their number
-            profiler: the profiler class to use
-            profiler_configuration: Optional profiler configuration dict
-            dry_run: when true, the method checks arguments and reports if can profile or specifies the arguments that are missing
-            additional_batch_kwargs: Additional keyword arguments to be provided to get_batch when loading the data asset.
-        Returns:
-            A dictionary::
-
-                {
-                    "success": True/False,
-                    "results": List of (expectation_suite, EVR) tuples for each of the data_assets found in the datasource
-                }
-
-            When success = False, the error details are under "error" key
-        """
-
-        # We don't need the datasource object, but this line serves to check if the datasource by the name passed as
-        # an arg exists and raise an error if it does not.
-        datasource = self.get_datasource(datasource_name)
-        assert datasource
-
-        if not dry_run:
-            logger.info(f"Profiling '{datasource_name}' with '{profiler.__name__}'")
-
-        profiling_results = {}
-
-        # Build the list of available data asset names (each item a tuple of name and type)
-
-        data_asset_names_dict = self.get_available_data_asset_names(datasource_name)
-
-        available_data_asset_name_list = []
-        try:
-            datasource_data_asset_names_dict = data_asset_names_dict[datasource_name]
-        except KeyError:
-            # KeyError will happen if there is not datasource
-            raise ge_exceptions.ProfilerError(f"No datasource {datasource_name} found.")
-
-        if batch_kwargs_generator_name is None:
-            # if no generator name is passed as an arg and the datasource has only
-            # one generator with data asset names, use it.
-            # if ambiguous, raise an exception
-            for name in datasource_data_asset_names_dict.keys():
-                if batch_kwargs_generator_name is not None:
-                    profiling_results = {
-                        "success": False,
-                        "error": {
-                            "code": BaseDataContext.PROFILING_ERROR_CODE_MULTIPLE_BATCH_KWARGS_GENERATORS_FOUND
-                        },
-                    }
-                    return profiling_results
-
-                if len(datasource_data_asset_names_dict[name]["names"]) > 0:
-                    available_data_asset_name_list = datasource_data_asset_names_dict[
-                        name
-                    ]["names"]
-                    batch_kwargs_generator_name = name
-
-            if batch_kwargs_generator_name is None:
-                profiling_results = {
-                    "success": False,
-                    "error": {
-                        "code": BaseDataContext.PROFILING_ERROR_CODE_NO_BATCH_KWARGS_GENERATORS_FOUND
-                    },
-                }
-                return profiling_results
-        else:
-            # if the generator name is passed as an arg, get this generator's available data asset names
-            try:
-                available_data_asset_name_list = datasource_data_asset_names_dict[
-                    batch_kwargs_generator_name
-                ]["names"]
-            except KeyError:
-                raise ge_exceptions.ProfilerError(
-                    "batch kwargs Generator {} not found. Specify the name of a generator configured in this datasource".format(
-                        batch_kwargs_generator_name
-                    )
-                )
-
-        available_data_asset_name_list = sorted(
-            available_data_asset_name_list, key=lambda x: x[0]
-        )
-
-        if len(available_data_asset_name_list) == 0:
-            raise ge_exceptions.ProfilerError(
-                "No Data Assets found in Datasource {}. Used batch kwargs generator: {}.".format(
-                    datasource_name, batch_kwargs_generator_name
-                )
-            )
-        total_data_assets = len(available_data_asset_name_list)
-
-        if isinstance(data_assets, list) and len(data_assets) > 0:
-            not_found_data_assets = [
-                name
-                for name in data_assets
-                if name not in [da[0] for da in available_data_asset_name_list]
-            ]
-            if len(not_found_data_assets) > 0:
-                profiling_results = {
-                    "success": False,
-                    "error": {
-                        "code": BaseDataContext.PROFILING_ERROR_CODE_SPECIFIED_DATA_ASSETS_NOT_FOUND,
-                        "not_found_data_assets": not_found_data_assets,
-                        "data_assets": available_data_asset_name_list,
-                    },
-                }
-                return profiling_results
-
-            data_assets.sort()
-            data_asset_names_to_profiled = data_assets
-            total_data_assets = len(available_data_asset_name_list)
-            if not dry_run:
-                logger.info(
-                    f"Profiling the white-listed data assets: {','.join(data_assets)}, alphabetically."
-                )
-        else:
-            if not profile_all_data_assets:
-                if total_data_assets > max_data_assets:
-                    profiling_results = {
-                        "success": False,
-                        "error": {
-                            "code": BaseDataContext.PROFILING_ERROR_CODE_TOO_MANY_DATA_ASSETS,
-                            "num_data_assets": total_data_assets,
-                            "data_assets": available_data_asset_name_list,
-                        },
-                    }
-                    return profiling_results
-
-            data_asset_names_to_profiled = [
-                name[0] for name in available_data_asset_name_list
-            ]
-        if not dry_run:
-            logger.info(
-                f"Profiling all {len(available_data_asset_name_list)} data assets from batch kwargs generator {batch_kwargs_generator_name}"
-            )
-        else:
-            logger.info(
-                f"Found {len(available_data_asset_name_list)} data assets from batch kwargs generator {batch_kwargs_generator_name}"
-            )
-
-        profiling_results["success"] = True
-
-        if not dry_run:
-            profiling_results["results"] = []
-            total_columns, total_expectations, total_rows, skipped_data_assets = (
-                0,
-                0,
-                0,
-                0,
-            )
-            total_start_time = datetime.datetime.now()
-
-            for name in data_asset_names_to_profiled:
-                logger.info(f"\tProfiling '{name}'...")
-                try:
-                    profiling_results["results"].append(
-                        self.profile_data_asset(
-                            datasource_name=datasource_name,
-                            batch_kwargs_generator_name=batch_kwargs_generator_name,
-                            data_asset_name=name,
-                            profiler=profiler,
-                            profiler_configuration=profiler_configuration,
-                            run_id=run_id,
-                            additional_batch_kwargs=additional_batch_kwargs,
-                            run_name=run_name,
-                            run_time=run_time,
-                        )["results"][0]
-                    )
-
-                except ge_exceptions.ProfilerError as err:
-                    logger.warning(err.message)
-                except OSError as err:
-                    logger.warning(
-                        f"IOError while profiling {name[1]}. (Perhaps a loading error?) Skipping."
-                    )
-                    logger.debug(str(err))
-                    skipped_data_assets += 1
-                except SQLAlchemyError as e:
-                    logger.warning(
-                        f"SqlAlchemyError while profiling {name[1]}. Skipping."
-                    )
-                    logger.debug(str(e))
-                    skipped_data_assets += 1
-
-            total_duration = (
-                datetime.datetime.now() - total_start_time
-            ).total_seconds()
-            logger.info(
-                f"""
-    Profiled {len(data_asset_names_to_profiled)} of {total_data_assets} named data assets, with {total_rows} total rows and {total_columns} columns in {total_duration:.2f} seconds.
-    Generated, evaluated, and stored {total_expectations} Expectations during profiling. Please review results using data-docs."""
-            )
-            if skipped_data_assets > 0:
-                logger.warning(
-                    f"Skipped {skipped_data_assets} data assets due to errors."
-                )
-
-        profiling_results["success"] = True
-        return profiling_results
 
     def list_checkpoints(self) -> Union[List[str], List[ConfigurationIdentifier]]:
         return self.checkpoint_store.list_checkpoints(ge_cloud_mode=self.ge_cloud_mode)
